requirements.txt
langchain
langchain_groq
langchain_community
langgraph
python-dotenv
langchain_openai
pydantic
langchain-anthropic
wikipedia 
uvicorn
fastapi
streamlit